Use JSON for tag and feedback requests.
I switched tag and feedback requests to use JSON rather than form data.
(The API supports both.) This simplifies request construction and makes
it easier to add more request parameters later.

This is a breaking change for the Tag function, but it should help
avoid more breaking changes in the future.